Add files via upload
The following is a project I had done as a part of my 6th semester project.The project is a customer cross selling upselling model performed on the sample superstore dataset which is available on Kaggle. Cross-selling involves selling related, supplementary products or services based on the customer's interest in, or purchase of, one of your company's products. Upselling is a sales strategy that involves encouraging customers to buy a higher-end version of a product than what they originally intended to purchase.The main aim of my project was to implement various marketing strategies and help boost sales.